Allow gateway to be driven directly using declarations
- Rejig the order of commands so that the root is now:
  - apiman gateway ...
  - apiman manager ...

Shares the same fundamental data model as manager; with a small tweak to policy which is non-breaking for the manager functionality.

For both inbuilt and plugin policies it resolves the correct information needed to successfully interact with the Gateway API. This should substantially simplify things.

Plugins are resolved using the Manager libraries, so it should function near identically.